+++ src/Core/Tests/DataSource/DataSourceManagerTest.php
@@ -46,7 +46,7 @@ class DataSourceManagerTest extends \PHPUnit_Framework_TestCase
 
         $dsm->load();
 
-        $this->assertCount(12, $dsm->getItems());
+        $this->assertCount(13, $dsm->getItems());
         $this->assertCount(1, $dsm->getLayouts());
         $this->assertCount(1, $dsm->getIncludes());
     }
@@ -63,7 +63,7 @@ class DataSourceManagerTest extends \PHPUnit_Framework_TestCase
         $dsm->load();
         $dsm->load();
 
-        $this->assertCount(12, $dsm->getItems());
+        $this->assertCount(13, $dsm->getItems());
         $this->assertCount(1, $dsm->getLayouts());
         $this->assertCount(1, $dsm->getIncludes());
     }
@@ -90,7 +90,7 @@ class DataSourceManagerTest extends \PHPUnit_Framework_TestCase
         $this->assertContains('filesystem_2', $dsm->getDataSourceNames());
         $this->assertNotContains('filesystem_1', $dsm->getDataSourceNames());
 
-        $this->assertCount(12, $dsm->getItems());
+        $this->assertCount(13, $dsm->getItems());
         $this->assertCount(1, $dsm->getLayouts());
         $this->assertCount(1, $dsm->getIncludes());
     }