+++ tests/Command/SiteBuildCommandTest.php
@@ -49,7 +49,7 @@ class SiteBuildCommandTest extends \PHPUnit_Framework_TestCase
         $this->assertRegExp('/Environment: dev/', $output);
         $this->assertNotRegExp('/Posts drafts enabled/', $output);
         $this->assertRegExp('/Debug mode: enabled/', $output);
-        $this->assertRegExp('/Total items: 16/', $output);
+        $this->assertRegExp('/Total items: 18/', $output);
     }
 
     public function testBuildCommandDraft()
@@ -69,7 +69,7 @@ class SiteBuildCommandTest extends \PHPUnit_Framework_TestCase
         $this->assertRegExp('/Starting.../', $output);
         $this->assertRegExp('/Environment: dev/', $output);
         $this->assertRegExp('/Draft posts: enabled/', $output);
-        $this->assertRegExp('/Total items: 16/', $output);
+        $this->assertRegExp('/Total items: 18/', $output);
     }
 
     public function testBuildCommandSafe()