+++ src/Plugin/Environment/CommandNotFoundException.php
@@ -20,7 +20,7 @@ class CommandNotFoundException extends \LogicException
 {
     /**
      * Constructor.
-     * 
+     *
      * @param string    $message      Exception message to throw.
      * @param int       $code         Exception code.
      * @param Exception $previous     previous exception used for the exception chaining.