+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 CHANGELOG for 2.x
 =================
 ## 2.1.2
+* [New] Two new methods have been added to StringWrapper class: `lower` and `upper`.
 * [Fixed] Bug #80 has been fixed: "Composer install fails if no namespace is specified in new plugin". See PR #81.
 * [Fixed] Bug #82 has been fixed: "Substr "---" in yaml string cause builder crash".
 * [Fixed] Bug #83 has been fixed: "Same tags on different languages cause builder crash".